requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
sqlalchemy 
pymysql 
python-dotenv
